+++ include/net/sock.h
@@ -434,7 +434,7 @@ struct sock {
 #endif
 	u16			sk_tsflags;
 	u8			sk_shutdown;
-	__u8			sk_ndm_mark;
+	u32			sk_ndm_mark;
 	u32			sk_tskey;
 	struct socket		*sk_socket;
 	void			*sk_user_data;
